chat application fixtures to enable the docker image to load successfully, fixing circular depedency
