+++ app/app/utils.py
@@ -462,7 +462,7 @@ def get_profile(request):
     if profile.chat_id is '' or profile.gitcoin_chat_access_token is '':
 
         try:
-            from chat.utils import associate_chat_to_profile
+            from chat.tasks import associate_chat_to_profile
             created, profile = associate_chat_to_profile(profile)
 
         except Exception as e:
