+++ app/chat/fixtures/initial.json
@@ -1,0 +1,17 @@
+[
+  {
+  "model": "oauth2.provider_accesstoken",
+  "pk": 1,
+  "fields": {
+    "created": "2020-02-29 21:30:30.418260 +00:00",
+    "updated": "2020-02-29 21:30:30.418279 +00:00",
+    "client_id": "VkHFfFlxnVCMYtYCWRxphbvPQUJjmER4cWtwqe7Q",
+    "client_secret": "xhLXNqXnOlkUiGCDFJLxQJjPb3CltLZNFZtzOsLgFfvwbYY60kYWRRBmqyjzfrnNyT6WWdIbfQEPpU854K2UiofmXtybFD1QkfIR9VVQyK3ttJpET7GormIBjK9CQRap",
+    "redirect_uris": "http://localhost:8065/gitcoin/login/complete http://localhost:8065/gitcoin/signup/complete http://localhost:8065/gitcoin/oauth/complete",
+    "client_type": "confidential",
+    "authorization_grant_type": "authorization-code",
+    "skip_authorization": true,
+    "user_id": null,
+    "name": "Gitcoin Chat"
+  }
+}]
