+++ app/chat/tasks.py
@@ -1,9 +1,11 @@
 from app.redis_service import RedisService
 from celery import app, group
 from celery.utils.log import get_task_logger
-from chat.utils import create_channel_if_not_exists, associate_chat_to_profile
 from dashboard.models import HackathonEvent, HackathonRegistration, Bounty, Profile
 import datetime
+import logging
+
+logger = logging.getLogger(__name__)
 from django.conf import settings
 from django.utils.text import slugify
 
@@ -25,6 +27,91 @@ driver_opts = {
 }
 
 chat_driver = Driver(driver_opts)
+
+
+def create_channel_if_not_exists(channel_opts):
+    try:
+        chat_driver = get_driver()
+
+        channel_lookup_response = chat_driver.channels.get_channel_by_name(
+            channel_opts['team_id'], channel_opts['channel_name']
+        )
+        return False, channel_lookup_response
+
+    except ResourceNotFound as RNF:
+        try:
+            result = create_channel.apply_async(args=[channel_opts])
+            channel_id_response = result.get()
+
+            if 'message' in channel_id_response:
+                raise ValueError(channel_id_response['message'])
+
+            return True, channel_id_response
+        except Exception as e:
+            logger.error(str(e))
+
+
+def update_chat_notifications(profile, notification_key, status):
+    query_opts = {}
+    if profile.chat_id is not '' or profile.chat_id is not None:
+        query_opts['chat_id'] = profile.chat_id
+    else:
+        query_opts['handle'] = profile.handle
+
+    update_user.delay(query_opts=query_opts, update_opts={'notify_props': {notification_key: status}})
+
+
+def associate_chat_to_profile(profile):
+    if profile.chat_id is not '':
+        return False, profile
+
+    chat_driver = get_driver()
+    try:
+
+        current_chat_user = chat_driver.users.get_user_by_username(profile.handle)
+        profile.chat_id = current_chat_user['id']
+        profile.save()
+
+        if profile.gitcoin_chat_access_token is not '':
+            profile_access_token = chat_driver.users.get_user_access_token(profile.gitcoin_chat_access_token)
+            profile.gitcoin_chat_access_token = profile_access_token['id']
+        return False, current_chat_user
+    except ResourceNotFound as RNF:
+        if profile.chat_id is '':
+            new_user_request = create_user.apply_async(args=[{
+                "email": profile.user.email,
+                "username": profile.handle,
+                "first_name": profile.user.first_name,
+                "last_name": profile.user.last_name,
+                "nickname": profile.handle,
+                "auth_data": f'{profile.user.id}',
+                "auth_service": "gitcoin",
+                "locale": "en",
+                "props": {},
+                "notify_props": {
+                    "email": "false",
+                    "push": "mention",
+                    "desktop": "all",
+                    "desktop_sound": "true",
+                    "mention_keys": f'{profile.handle}, @{profile.handle}',
+                    "channel": "true",
+                    "first_name": "false"
+                },
+            }, {
+                "tid": settings.GITCOIN_HACK_CHAT_TEAM_ID
+            }])
+            response = new_user_request.get()
+            profile.chat_id = response['id']
+            profile.save()
+        else:
+            try:
+                profile_access_token = chat_driver.users.create_user_access_token(user_id=profile.chat_id, options={
+                    'description': "Grants Gitcoin access to modify your account"})
+                profile.gitcoin_chat_access_token = profile_access_token['id']
+            except Exception as e:
+                logger.error(str(e))
+
+        return True, profile
 
 
 def get_chat_url():
@@ -65,9 +152,12 @@ def create_channel(self, options, bounty_id=None, retry: bool = True) -> None:
             new_channel = chat_driver.channels.create_channel(options={
                 'team_id': options['team_id'],
                 'name': options['channel_name'],
+                'purpose': options['channel_purpose'] if options['channel_purpose'] else '',
+                'header': options['channel_header'] if options['channel_header'] else '',
                 'display_name': options['channel_display_name'],
-                'type': 'O'
+                'type': options['channel_type'] or 'O'
             })
+
             if bounty_id:
                 active_bounty = Bounty.objects.get(id=bounty_id)
                 active_bounty.chat_channel_id = new_channel['id']
@@ -84,7 +174,6 @@ def create_channel(self, options, bounty_id=None, retry: bool = True) -> None:
 
 @app.shared_task(bind=True, max_retries=3)
 def hackathon_chat_sync(self, hackathon_id: str, profile_handle: str = None, retry: bool = True) -> None:
-    tasks = []
     try:
         hackathon = HackathonEvent.objects.get(id=hackathon_id)
         channels_to_connect = []
